coord: split explain work into its own file
