+++ src/dataflow-types/src/explain.rs
@@ -231,44 +231,3 @@ impl<'a> ViewFormatter<OptimizedMirRelationExpr> for DataflowGraphFormatter<'a> 
         fmt::Display::fmt(&explain, f)
     }
 }
-
-/// Information used when determining the timestamp for a query.
-pub struct TimestampExplanation<T> {
-    /// The chosen timestamp from `determine_timestamp`.
-    pub timestamp: T,
-    /// Whether the query contains a table.
-    pub has_table: bool,
-    /// If the query contains a table, the global table read timestamp.
-    pub table_read_ts: Option<T>,
-    /// The read frontier of all involved sources.
-    pub since: Vec<T>,
-    /// The write frontier of all involved sources.
-    pub upper: Vec<T>,
-    /// Details about each source.
-    pub sources: Vec<TimestampSource<T>>,
-}
-
-pub struct TimestampSource<T> {
-    pub name: String,
-    pub read_frontier: Vec<T>,
-    pub write_frontier: Vec<T>,
-}
-
-impl<T: fmt::Display + fmt::Debug> fmt::Display for TimestampExplanation<T> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        writeln!(f, "     timestamp: {}", self.timestamp)?;
-        writeln!(f, "         since:{:13?}", self.since)?;
-        writeln!(f, "         upper:{:13?}", self.upper)?;
-        writeln!(f, "     has table: {}", self.has_table)?;
-        if let Some(ts) = &self.table_read_ts {
-            writeln!(f, " table read ts: {ts}")?;
-        }
-        for source in &self.sources {
-            writeln!(f, "")?;
-            writeln!(f, "source {}:", source.name)?;
-            writeln!(f, " read frontier:{:13?}", source.read_frontier)?;
-            writeln!(f, "write frontier:{:13?}", source.write_frontier)?;
-        }
-        Ok(())
-    }
-}
