+++ src/dataflow-types/src/lib.rs
@@ -26,8 +26,6 @@ pub use errors::*;
 pub use explain::DataflowGraphFormatter;
 pub use explain::Explanation;
 pub use explain::JsonViewFormatter;
-pub use explain::TimestampExplanation;
-pub use explain::TimestampSource;
 pub use gen::*;
 pub use plan::Plan;
 pub use types::*;
